+++ src/js/others.js
@@ -15,6 +15,7 @@ import {
   forEach,
   getMaxZoomRatio,
   isFunction,
+  isPlainObject,
   isString,
   removeClass,
   removeListener,
@@ -138,7 +139,7 @@ export default {
     }
   },
 
-  requestFullscreen() {
+  requestFullscreen(options) {
     const document = this.element.ownerDocument;
 
     if (this.fulled && !(
@@ -151,7 +152,12 @@ export default {
 
       // Element.requestFullscreen()
       if (documentElement.requestFullscreen) {
-        documentElement.requestFullscreen();
+        // Avoid TypeError when convert `options` to dictionary
+        if (isPlainObject(options)) {
+          documentElement.requestFullscreen(options);
+        } else {
+          documentElement.requestFullscreen();
+        }
       } else if (documentElement.webkitRequestFullscreen) {
         documentElement.webkitRequestFullscreen(Element.ALLOW_KEYBOARD_INPUT);
       } else if (documentElement.mozRequestFullScreen) {