+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next
 
+- Enhance `fullscreen` option and `play` method to support [`FullscreenOptions`](https://developer.mozilla.org/en-US/docs/Web/API/FullscreenOptions) (#482, #483).
 - Refactor the `toggle` method for toggling image between current/natural ratio (#477).
 
 ## 1.9.2 (May 29, 2021)